SAIL Plugins/Platform Monitoring/requirements.txt
pandas
watchdog
azure-identity
azure-mgmt-resource
azure-mgmt-network
azure-mgmt-compute
azure-storage-blob
mysql-connector-python
requests
pywin32